added multiple args

+++ commands/enablerole.js
@@ -9,16 +9,33 @@ module.exports = {
 };
 
 module.exports.run = (client, message, args) => {
-  // Find the role in the guild
-  const role = message.guild.roles.find(elem => elem.name.toLowerCase() === args.toLowerCase());
+  // Support for multiple roles
+  const roles = args.split(',').map(arg => _.trim(arg)).filter(Boolean);
 
-  // Return if the role doesnt exist.
-  if (!role) {
-    message.channel.send('Please enter a valid role name.');
-    return;
-  }
+  roles.forEach((roleName) => {
+    // Find the role in the guild
+    const role = message.guild.roles.find(elem => elem.name.toLowerCase() === roleName.toLowerCase());
 
-  // Enable the role
-  db.removeDisabledRole(message.guild, role);
-  message.react('✅');
+    // Return if the role doesnt exist.
+    if (!role) {
+      message.react('❓');
+      message.channel.send(`"${roleName}" is not a valid role name.`).then((msg) => {
+        msg.delete(5000);
+      });
+      return;
+    }
+
+    // Check if the role isn't disabled
+    if (!db.roleIsDisabled(message.guild, role)) {
+      message.react('❌');
+      message.channel.send(`"${roleName}" is already enabled.`).then((msg) => {
+        msg.delete(5000);
+      });
+      return;
+    }
+
+    // Enable the role
+    db.removeDisabledRole(message.guild, role);
+    message.react('✅');
+  });
 };
